extract selectable
